apply pr feedback
