+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/datareaders/zarr3/Zarr3ArrayHeader.scala
@@ -251,7 +251,7 @@ object Zarr3ArrayHeader extends JsonImplicits {
       )
 
   }
-  def fromDataLayerToVersion3(dataLayer: DataLayer): Zarr3ArrayHeader = {
+  def fromDataLayer (dataLayer: DataLayer): Zarr3ArrayHeader = {
     val additionalAxes = reorderAdditionalAxes(dataLayer.additionalAxes.getOrElse(Seq.empty))
     Zarr3ArrayHeader(
       zarr_format = 3,
